back to vs compiler with cmake fix some warnings server and lobby build

+++ src/effect/alephgeo.cpp
@@ -131,7 +131,7 @@ public:
             int         index = indexDepth * (m_countSection + 1);
             float       value = (float)indexDepth / m_countDepth;
             float           z = time + value;
-            float angleOffset = 5 * (1 - pow(1 - value, 2));
+            float angleOffset = 5 * (1 - (float)pow(1 - value, 2));
 
             for (int indexSection = 0; indexSection < (m_countSection + 1); indexSection++) {
                 float angle = 4 * (float)indexSection / m_countSection + angleOffset;
